requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
cryptography==42.0.2
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
numpy==1.25.1
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.2.0
pandas==2.0.3
Pillow==10.2.0
psycopg2==2.9.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.15.2
six==1.16.0
slack-sdk==3.24.0
slackclient==2.9.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
thefuzz==0.22.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.2
zstandard==0.21.0  
